inference/requirements.txt
# tensorflow>=2.13.0
matplotlib>=3.5.0
numpy>=1.21.0
tqdm>=4.64.0
tensorflow-addons>=0.20.0  

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
requests>=2.31.0
pydantic>=2.0.0
